backend/requirements.txt
CopyFlask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-SocketIO==5.1.1
Flask-JWT-Extended==4.3.1
Flask-Limiter==1.4
psycopg2-binary==2.9.1
celery==5.1.2
redis==3.5.3
pytest==6.2.5
gunicorn==20.1.0

